+++ src/rust/lib/flexer/src/automata.rs
@@ -1,3 +1,5 @@
+//! Exports API for construction of Nondeterminist and Deterministic Finite State Automata.
+
 pub mod alphabet;
 pub mod dfa;
 pub mod nfa;
