+++ src/rust/lib/flexer/src/automata/alphabet.rs
@@ -1,3 +1,5 @@
+//! Exports an alphabet (set of all valid input symbols) for Finite State Automata (NFA and DFA).
+
 use crate::automata::state::Symbol;
 
 use std::collections::BTreeSet;
@@ -9,19 +11,17 @@ use std::ops::RangeInclusive;
 // === Alphabet ===
 // ================
 
-/// An alphabet describes a set of all the valid input symbols
-/// that a given finite state automata (NFA or DFA) can operate over.
+/// An alphabet describes a set of all the valid input symbols that a given finite state automata
+/// (NFA or DFA) can operate over.
 /// More information at: https://en.wikipedia.org/wiki/Deterministic_finite_automaton
-/// The alphabet is meant to be represented as an interval.
-/// That is, if `a` and `b` are in alphabet,
+/// The alphabet is meant to be represented as an interval. That is, if `a` and `b` are in alphabet,
 /// then any symbol from `a..=b` is in alphabet too.
 #[derive(Clone,Debug,PartialEq,Eq)]
 pub struct Alphabet {
-    /// The interval of all valid input symbols.
-    /// The interval is further divided into subintervals
-    /// (i.e. `[a,z,A,Z]` should be understood as `[a..=z,z..=A,A..=Z]`),
-    /// in order to efficiently encode state transitions
-    /// that trigger not just on one but a whole range of symbols (i.e. `a..=z`)
+    /// The interval of all valid input symbols. The interval is further divided into subintervals
+    /// (i.e. `[a,z,A,Z]` should be understood as `[a..=z,z..=A,A..=Z]`), in order to efficiently
+    /// encode state transitions that trigger not just on one but a whole range of symbols
+    /// (i.e. `a..=z`)
     pub symbols: BTreeSet<Symbol>
 }
 
@@ -34,8 +34,12 @@ impl Default for Alphabet {
 impl Alphabet {
     /// Inserts a range of symbols into the alphabet.
     pub fn insert(&mut self, range:RangeInclusive<Symbol>) {
+        // The symbol range is associated with transition in automata. Therefore we:
+        // Mark the symbol with the new transition.
         self.symbols.insert(Symbol{val:range.start().val});
+        // Mark the symbol without the new transition.
         self.symbols.insert(Symbol{val:range.end().val + 1});
+        // This way each symbol in alphabet corresponds to a unique set of transitions.
     }
 }
