+++ src/rust/lib/flexer/src/data.rs
@@ -1,0 +1,3 @@
+//! Generic datastructures, with multiple usecases.
+
+pub mod matrix;
